+++ include/nfx/silicon/Decoder.h
@@ -5,6 +5,7 @@
  * \brief Convenience header for all decoder packages.
  */
 
+#include "decoder/cd4xxx/CD4028B.h"
 #include "decoder/cd4xxx/CD4556B.h"
 
 #include "decoder/sn74xxx/LS138.h"